chore: create user dialog package
