+++ packages/paste-codemods/tools/.cache/mappings.json
@@ -279,18 +279,18 @@
   "Tooltip": "@twilio-paste/core/tooltip",
   "useTooltipState": "@twilio-paste/core/tooltip",
   "Topbar": "@twilio-paste/core/topbar",
-  "UserDialog": "@twilio-paste/core/topbar",
-  "UserDialogContainer": "@twilio-paste/core/topbar",
-  "UserDialogContext": "@twilio-paste/core/topbar",
-  "UserDialogList": "@twilio-paste/core/topbar",
-  "UserDialogListItem": "@twilio-paste/core/topbar",
-  "UserDialogSeparator": "@twilio-paste/core/topbar",
-  "UserDialogUserEmail": "@twilio-paste/core/topbar",
-  "UserDialogUserInfo": "@twilio-paste/core/topbar",
-  "UserDialogUserName": "@twilio-paste/core/topbar",
-  "useUserDialogListState": "@twilio-paste/core/topbar",
-  "useUserDialogState": "@twilio-paste/core/topbar",
   "Truncate": "@twilio-paste/core/truncate",
+  "UserDialog": "@twilio-paste/core/user-dialog",
+  "UserDialogContainer": "@twilio-paste/core/user-dialog",
+  "UserDialogContext": "@twilio-paste/core/user-dialog",
+  "UserDialogList": "@twilio-paste/core/user-dialog",
+  "UserDialogListItem": "@twilio-paste/core/user-dialog",
+  "UserDialogSeparator": "@twilio-paste/core/user-dialog",
+  "UserDialogUserEmail": "@twilio-paste/core/user-dialog",
+  "UserDialogUserInfo": "@twilio-paste/core/user-dialog",
+  "UserDialogUserName": "@twilio-paste/core/user-dialog",
+  "useUserDialogListState": "@twilio-paste/core/user-dialog",
+  "useUserDialogState": "@twilio-paste/core/user-dialog",
   "AspectRatio": "@twilio-paste/core/aspect-ratio",
   "Flex": "@twilio-paste/core/flex",
   "Column": "@twilio-paste/core/grid",
