+++ packages/paste-core/components/topbar/src/index.tsx
@@ -1,11 +1,1 @@
-/* Wrapper */
 export * from './Topbar';
-/* User Dialog */
-export * from './user-dialog/UserDialog';
-export * from './user-dialog/UserDialogContainer';
-export * from './user-dialog/UserDialogUserInfo';
-export * from './user-dialog/UserDialogUserName';
-export * from './user-dialog/UserDialogUserEmail';
-export * from './user-dialog/UserDialogList';
-export * from './user-dialog/UserDialogListItem';
-export * from './user-dialog/UserDialogSeparator';
