+++ packages/paste-core/components/user-dialog/package.json
@@ -1,0 +1,84 @@
+{
+  "name": "@twilio-paste/user-dialog",
+  "version": "0.0.0",
+  "category": "navigation",
+  "status": "production",
+  "description": "User dialog is an expandable menu that contains profile related actions.",
+  "author": "Twilio Inc.",
+  "license": "MIT",
+  "main:dev": "src/index.tsx",
+  "main": "dist/index.js",
+  "module": "dist/index.es.js",
+  "types": "dist/index.d.ts",
+  "sideEffects": false,
+  "publishConfig": {
+    "access": "public"
+  },
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "build": "yarn clean && NODE_ENV=production node build.js && tsc",
+    "build:js": "NODE_ENV=development node build.js",
+    "clean": "rm -rf ./dist",
+    "tsc": "tsc"
+  },
+  "peerDependencies": {
+    "@twilio-paste/anchor": "^11.0.0",
+    "@twilio-paste/animation-library": "^1.0.0",
+    "@twilio-paste/avatar": "^8.0.0",
+    "@twilio-paste/box": "^9.0.0",
+    "@twilio-paste/button": "^13.0.1",
+    "@twilio-paste/color-contrast-utils": "^4.0.0",
+    "@twilio-paste/customization": "^7.0.0",
+    "@twilio-paste/design-tokens": "^9.0.0",
+    "@twilio-paste/icons": "^11.0.0",
+    "@twilio-paste/listbox-primitive": "^2.0.0",
+    "@twilio-paste/media-object": "^9.0.0",
+    "@twilio-paste/non-modal-dialog-primitive": "^1.0.0",
+    "@twilio-paste/reakit-library": "^1.0.0",
+    "@twilio-paste/separator": "^7.0.0",
+    "@twilio-paste/spinner": "^13.0.0",
+    "@twilio-paste/stack": "^7.0.0",
+    "@twilio-paste/style-props": "^8.0.0",
+    "@twilio-paste/styling-library": "^2.0.0",
+    "@twilio-paste/text": "^9.0.1",
+    "@twilio-paste/theme": "^10.0.0",
+    "@twilio-paste/types": "^5.0.0",
+    "@twilio-paste/uid-library": "^1.0.0",
+    "@types/react": "^16.8.6 || ^17.0.2 || ^18.0.27",
+    "@types/react-dom": "^16.8.6 || ^17.0.2 || ^18.0.10",
+    "prop-types": "^15.7.2",
+    "react": "^16.8.6 || ^17.0.2 || ^18.0.0",
+    "react-dom": "^16.8.6 || ^17.0.2 || ^18.0.0"
+  },
+  "devDependencies": {
+    "@twilio-paste/anchor": "^11.0.0",
+    "@twilio-paste/animation-library": "^1.0.0",
+    "@twilio-paste/avatar": "^8.1.0",
+    "@twilio-paste/box": "^9.0.0",
+    "@twilio-paste/button": "^13.0.2",
+    "@twilio-paste/color-contrast-utils": "^4.0.0",
+    "@twilio-paste/customization": "^7.0.0",
+    "@twilio-paste/design-tokens": "^9.2.0",
+    "@twilio-paste/icons": "^11.1.0",
+    "@twilio-paste/listbox-primitive": "^2.0.0",
+    "@twilio-paste/media-object": "^9.0.0",
+    "@twilio-paste/non-modal-dialog-primitive": "^1.0.0",
+    "@twilio-paste/reakit-library": "^1.0.0",
+    "@twilio-paste/separator": "^7.0.0",
+    "@twilio-paste/spinner": "^13.0.0",
+    "@twilio-paste/stack": "^7.0.0",
+    "@twilio-paste/style-props": "^8.0.0",
+    "@twilio-paste/styling-library": "^2.0.0",
+    "@twilio-paste/text": "^9.0.1",
+    "@twilio-paste/theme": "^10.0.0",
+    "@twilio-paste/types": "^5.0.0",
+    "@twilio-paste/uid-library": "^1.0.0",
+    "@types/react": "^18.0.27",
+    "@types/react-dom": "^18.0.10",
+    "prop-types": "^15.7.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0"
+  }
+}
