+++ packages/paste-core/components/user-dialog/src/index.tsx
@@ -1,0 +1,8 @@
+export * from './UserDialog';
+export * from './UserDialogContainer';
+export * from './UserDialogUserInfo';
+export * from './UserDialogUserName';
+export * from './UserDialogUserEmail';
+export * from './UserDialogList';
+export * from './UserDialogListItem';
+export * from './UserDialogSeparator';
