+++ packages/paste-core/core-bundle/.gitignore
@@ -95,4 +95,5 @@
 /truncate
 /types
 /uid-library
+/user-dialog
 /utils
