+++ packages/paste-core/core-bundle/src/index.tsx
@@ -77,3 +77,4 @@ export * from '@twilio-paste/tooltip';
 export * from '@twilio-paste/tooltip-primitive';
 export * from '@twilio-paste/topbar';
 export * from '@twilio-paste/truncate';
+export * from '@twilio-paste/user-dialog';
