+++ packages/paste-core/core-bundle/src/user-dialog.tsx
@@ -1,0 +1,1 @@
+export * from '@twilio-paste/user-dialog';
